refactor: split delete-all-data button and modal into separate components
Extract the confirmation modal into DeleteAllDataModal, leaving
DeleteAllDataButton as the trigger that owns the disclosure state, mirroring
the existing LogoutModal pattern. Mounting the modal only while open resets
the consent checkbox automatically, so the manual reset is gone.
